This filter is a video filter, not a frame filter
